Security

+++ src/main/java/com/codeup/springblog/controllers/PostController.java
@@ -4,6 +4,9 @@ import com.codeup.springblog.Repositories.PostRepository;
 import com.codeup.springblog.Repositories.UserRepository;
 import com.codeup.springblog.models.Post;
 import com.codeup.springblog.models.User;
+import com.codeup.springblog.services.EmailService;
+import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.security.core.context.SecurityContextHolder;
 import org.springframework.stereotype.Controller;
 import org.springframework.ui.Model;
 import org.springframework.web.bind.annotation.*;
@@ -14,9 +17,13 @@ public class PostController {
     private PostRepository postsDao;
     private UserRepository usersDao;
 
-    public PostController(PostRepository postsDao, UserRepository usersDao) {
+    @Autowired
+    private final EmailService emailService;
+
+    public PostController(PostRepository postsDao, UserRepository usersDao, EmailService emailService) {
         this.postsDao = postsDao;
         this.usersDao = usersDao;
+        this.emailService = emailService;
     }
 
     @GetMapping("/posts")
@@ -48,13 +55,19 @@ public class PostController {
         newPost.setBody(body);
         newPost.setUser(usersDao.getOne(1L));
         postsDao.save(newPost);
+        String emailSubject = "Your new post: " + title;
+        String emailBody = "Recently posted: " + body;
+
+        emailService.prepareAndSend(newPost, emailSubject, emailBody);
         return "redirect:/posts";
     }
 
     @PostMapping("/posts/{id}/delete")
     public String deletePost(@PathVariable long id) {
-        // delete post
-        postsDao.deleteById(id);
+        User loggedInUser = (User) SecurityContextHolder.getContext().getAuthentication().getPrincipal();
+        if (loggedInUser.getId() == postsDao.getOne(id).getUser().getId())
+            // delete post
+            postsDao.deleteById(id);
         return "redirect:/posts";
     }
 
@@ -69,6 +82,7 @@ public class PostController {
     public String updatePost(@PathVariable long id, @ModelAttribute Post post) {
         // post object from database currently
         Post editedPost = postsDao.getOne(id);
+
         editedPost.setTitle(post.getTitle());
         editedPost.setBody(post.getBody());
         postsDao.save(editedPost);
